Only update new or updated document parts for functions (#2)
Previous implementation would always delete documentation parts and reupload new ones. This however was inefficient and a drain on APIGateway concurrency limits if deploying multiple serverless stacks in parallel.

This PR will properly query and determine if a documentation part exists and whether it has changed only then is it deleted. Otherwise if a documentation part exists and remains unchanged, it is not recreated.

This will substantially decrease the number of APIGateway update requests allowing for faster serverless deployments.